Create HSV Colors

This function creates colors of class HSV; a subclass of the virtual color-class class.

HSV(H, S, V, names)

Arguments

H, S, V
These arguments give the hue, saturation and value of the colors. The values can be provided in separate H, S and V vectors or in a three-column matrix passed as H.

names
A vector of names for the colors (by default the row names of H are used).

Value

An object of class HSV which inherits from class color.

Details

This function creates colors in the HSV color space which corresponds to the standard sRGB color space (IEC standard 61966). The hues should lie between between 0 and 360, and the saturations and values should lie between 0 and 1.

See also

Examples

# A rainbow of full-intensity hues
HSV(seq(0, 360, length = 13)[-13], 1, 1)

#>         H S V
#>  [1,]   0 1 1
#>  [2,]  30 1 1
#>  [3,]  60 1 1
#>  [4,]  90 1 1
#>  [5,] 120 1 1
#>  [6,] 150 1 1
#>  [7,] 180 1 1
#>  [8,] 210 1 1
#>  [9,] 240 1 1
#> [10,] 270 1 1
#> [11,] 300 1 1
#> [12,] 330 1 1