XYZ {colorspace}	R Documentation

Create XYZ Colors

Description

This function creates colors of class XYZ; a subclass of the virtual color-class class.

Usage

XYZ(X, Y, Z, names)

Arguments

X, Y, Z
these arguments give the X, Y and Z coordinates of the colors. The values can be provided in separate X, Y and Z vectors or in a three-column matrix passed as X.

names
A vector of names for the colors (by default the row names of X are used).

Details

The X, Y and Z values are the levels of the CIE primaries. These are scaled so that the luminance of the display white-point is 100. The white-point is taken to be D65, which means that its coordinates are 95.047, 100.000, 108.883.

Value

An object of class XYZ which inherits from class color.

Author(s)

Ross Ihaka

See Also

RGB, HSV, LAB, polarLAB, LUV, polarLUV.

Examples

## Generate white in XYZ space
XYZ(95.047, 100.000, 108.883)

[Package colorspace version 2.1-0 Index]