coords {colorspace}	R Documentation

Extract the numerical coordinates of a color

Description

This function returns a matrix with three columns which give the coordinates of a color in its natural color space.

Usage

coords(color)

Arguments

color
A color.

Value

A numeric matrix giving the coordinates of the color.

Author(s)

Ross Ihaka.

See Also

RGB, XYZ, LAB, polarLAB, LUV, polarLUV, mixcolor.

Examples

x <- RGB(1, 0, 0)
coords(as(x, "HSV"))

[Package colorspace version 1.2-4 Index]