Convert Hexadecimal Color Specifications to sRGB Objects

This function takes a vector of strings of the form "#RRGGBB" (hexadecimal color descriptions) into sRGB objects.

hex2RGB(x, gamma = FALSE)

Arguments

x
a vector of hexadecimal color descriptions.

gamma
Whether to apply gamma-correction.

Value

An sRGB object describing the colors.

Details

This function converts device-dependent color descriptions of the form "#RRGGBB" into sRGB color descriptions (linearized if gamma is TRUE). The alpha channel will be ignored if given ("#RRGGBBAA").

See also

Author

Ross Ihaka

Examples

hex2RGB(c("#FF0000","#00FF00", "#0000FF50"))

#>      R G B
#> [1,] 1 0 0
#> [2,] 0 1 0
#> [3,] 0 0 1