polarLAB {colorspace}	R Documentation

Create polarLAB Colors

Description

This function creates colors of class “polarLAB”; a subclass of the virtual color-class class.

Usage

polarLAB(L, C, H, names)

Arguments

L, C, H
these arguments give the L, C and H coordinates of the colors. The values can be provided in separate L, C and H vectors or in a three-column matrix passed as L.

names
A vector of names for the colors (by default the row names of L are used).

Details

The polarLAB space is a transformation of the CIE L^*a^*b^* space so that the a and b values are converted to polar coordinates. The radial component C measures chroma and the angular coordinate H is measures hue.

Value

An object of class polarLAB which inherits from class color.

Author(s)

Ross Ihaka

See Also

RGB, HSV, XYZ, LAB, polarLAB, LUV, polarLUV.

Examples

## Show the polarLAB space
set.seed(1)
x <- sRGB(runif(1000), runif(1000), runif(1000))
y <- as(x, "polarLAB")
head(x)
head(y)
plot(y)

[Package colorspace version 2.1-0 Index]