HSV {colorspace}	R Documentation

Create HSV Colors

Description

This function creates colors of class HSV; a subclass of the virtual color-class class.

Usage

HSV(H, S, V, names)

Arguments

H, S, V
These arguments give the hue, saturation and value of the colors. The values can be provided in separate H, S and V vectors or in a three-column matrix passed as H.

names
A vector of names for the colors (by default the row names of H are used).

Details

This function creates colors in an HSV color space. The hues should lie between between 0 and 360, and the saturations and values should lie between 0 and 1.

HSV is a relative color space; it is a transformation of an RGB color space. Conversion of HSV colors to any other color space must first involve a conversion to a specific RGB color space, for example the standard sRGB color space (IEC standard 61966).

Value

An object of class HSV which inherits from class color.

Author(s)

Ross Ihaka

See Also

sRGB, RGB, XYZ, LAB, polarLAB, LUV, polarLUV.

Examples

# A rainbow of full-intensity hues
HSV(seq(0, 360, length.out = 13)[-13], 1, 1)

[Package colorspace version 2.1-0 Index]