RGB {colorspace}	R Documentation

Create RGB Colors

Description

This function creates colors of class RGB; a subclass of the virtual color-class class.

Usage

RGB(R, G, B, names)

Arguments

R, G, B
these arguments give the red, green and blue intensities of the colors (the values should lie between 0 and 1). The values can be provided in separate R, G and B vectors or in a three-column matrix passed as R.

names
A vector of names for the colors (by default the row names of R are used).

Details

This function creates colors in the linearized sRGB color space (IEC standard 61966).

Value

An object of class RGB which inherits from class color.

Author(s)

Ross Ihaka

See Also

sRGB, HSV, XYZ, LAB, polarLAB, LUV, polarLUV.

Examples

# Create a random set of colors
set.seed(1)
RGB(R = runif(20), G = runif(20), B = runif(20))

[Package colorspace version 2.0-2 Index]