coords {colorspace}	R Documentation

Extract the Numerical Coordinates of a Color

Description

This function returns a matrix with three columns which give the coordinates of a color in its natural color space.

Usage

coords(color)

Arguments

color
A color.

Value

A numeric matrix giving the coordinates of the color.

Author(s)

Ross Ihaka

See Also

RGB, XYZ, LAB, polarLAB, LUV, polarLUV, mixcolor.

Examples

x <- sRGB(1, 0, 0)
coords(as(x, "HSV"))

[Package colorspace version 2.1-0 Index]