cvd_emulator {colorspace}	R Documentation

Graphical User Interface to Check Images for Color Constraints

Description

A graphical user interface (GUI) to check an existing jpg/png image for (possible) color constraints. The image will be converted to protanope vision, deuteranope vision, and a desaturated version (monochromatic vision). Allows a rapid check whether the colors used in the image show some constraints with respect to color deficiency or color blindness.

Usage

cvd_emulator(file, overwrite = FALSE, shiny.trace = FALSE)

Arguments

file
If not set, an interactive GUI will be started. If x is of type character it has to be the full path to an image of type png or jpg/jpeg. The image will be converted and stored on disc, no GUI.

overwrite
logical. Only used if file is provided. Allow the function to overwrite files on disc if they exist.

shiny.trace
logical. Can be set to TRUE for more verbose output when the GUI is started (development flag).

Author(s)

Reto Stauffer, Claus O. Wilke, Achim Zeileis

References

Zeileis A, Fisher JC, Hornik K, Ihaka R, McWhite CD, Murrell P, Stauffer R, Wilke CO (2020). “colorspace: A Toolbox for Manipulating and Assessing Colors and Palettes.” Journal of Statistical Software, 96(1), 1–49. doi:10.18637/jss.v096.i01

[Package colorspace version 2.1-0 Index]