hcl_color_picker {colorspace}	R Documentation

Graphical User Interface to Pick Colors in HCL Space

Description

The app visualizes colors either along the hue-chroma plane for a given luminance value or along the luminance-chroma plane for a given hue. Colors can be entered by specifying the hue (H), chroma (C), and luminance (L) values via sliders, by entering an RGB hex code, or by clicking on a color in the hue-chroma or luminance-chroma plane. It is also possible to select individual colors and add them to a palette for comparison and future reference.

Usage

hcl_color_picker(shiny.trace = FALSE)

choose_color(shiny.trace = FALSE)

Arguments

shiny.trace
logical: used for debugging the shiny interface.

Details

choose_color is a convenience alias for hcl_color_picker to go along with choose_palette. Another alias is hclcolorpicker.

Value

hclcolorpicker invisibly returns a vector of colors choosen. If no colors have been selected NULL will be returned.

Author(s)

Claus O. Wilke, Reto Stauffer, Achim Zeileis

References

Zeileis A, Fisher JC, Hornik K, Ihaka R, McWhite CD, Murrell P, Stauffer R, Wilke CO (2020). “colorspace: A Toolbox for Manipulating and Assessing Colors and Palettes.” Journal of Statistical Software, 96(1), 1–49. doi: 10.18637/jss.v096.i01

See Also

choose_palette

Examples

## Not run: 
hcl_color_picker()

## End(Not run)

[Package colorspace version 2.0-2 Index]