Shiny App to Pick Colors in HCL Space

The app visualizes colors either along the hue-chroma plane for a given luminance value or along the luminance-chroma plane for a given hue. Colors can be entered by specifying the hue (H), chroma (C), and luminance (L) values via sliders, by entering an RGB hex code, or by clicking on a color in the hue-chroma or luminance-chroma plane. It is also possible to select individual colors and add them to a palette for comparison and future reference.

hcl_color_picker(shiny.trace = FALSE)

choose_color(shiny.trace = FALSE)

Arguments

shiny.trace
logical: used for debugging the shiny interface.

Value

hclcolorpicker invisibly returns a vector of colors choosen. If no colors have been selected NULL will be returned.

Details

choose_color is a convenience alias for hcl_color_picker to go along with choose_palette. Another alias is hclcolorpicker.

See also

Examples

# NOT RUN {
hcl_color_picker()
# }