hex2RGB {colorspace}	R Documentation

Convert Hexadecimal Color Specifications to sRGB Objects

Description

This function takes a vector of strings of the form "#RRGGBB" (hexadecimal color descriptions) into sRGB objects.

Usage

hex2RGB(x, gamma = FALSE)

Arguments

x
a vector of hexadecimal color descriptions.

gamma
Whether to apply gamma-correction.

Details

This function converts device-dependent color descriptions of the form "#RRGGBB" into sRGB color descriptions (linearized if gamma is TRUE). The alpha channel will be ignored if given ("#RRGGBBAA").

Value

An sRGB object describing the colors.

Author(s)

Ross Ihaka

See Also

hex, RGB, sRGB, HSV, XYZ, polarLAB, LUV, polarLUV.

Examples

hex2RGB(c("#FF0000","#00FF00", "#0000FF50"))

[Package colorspace version 2.0-2 Index]