scale_colour_binned_sequential {colorspace}	R Documentation

HCL-Based Binned Sequential Color Scales for ggplot2

Description

Binned ggplot2 color scales using the color palettes generated by sequential_hcl.

Usage

scale_colour_binned_sequential(
  palette = NULL,
  c1 = NULL,
  c2 = NULL,
  cmax = NULL,
  l1 = NULL,
  l2 = NULL,
  h1 = NULL,
  h2 = NULL,
  p1 = NULL,
  p2 = NULL,
  alpha = 1,
  rev = TRUE,
  begin = 0,
  end = 1,
  na.value = "grey50",
  guide = "coloursteps",
  aesthetics = "colour",
  n_interp = 11,
  ...
)

scale_color_binned_sequential(
  palette = NULL,
  c1 = NULL,
  c2 = NULL,
  cmax = NULL,
  l1 = NULL,
  l2 = NULL,
  h1 = NULL,
  h2 = NULL,
  p1 = NULL,
  p2 = NULL,
  alpha = 1,
  rev = TRUE,
  begin = 0,
  end = 1,
  na.value = "grey50",
  guide = "coloursteps",
  aesthetics = "colour",
  n_interp = 11,
  ...
)

scale_fill_binned_sequential(..., aesthetics = "fill")

Arguments

palette
The name of the palette to be used. Run hcl_palettes(type = "sequential") for available options.

c1
Beginning chroma value.

c2
Ending chroma value.

cmax
Maximum chroma value.

l1
Beginning luminance value.

l2
Ending luminance value.

h1
Beginning hue value.

h2
Ending hue value. If set to NA, generates a single-hue scale.

p1
Control parameter determining how chroma should vary (1 = linear, 2 = quadratic, etc.).

p2
Control parameter determining how luminance should vary (1 = linear, 2 = quadratic, etc.).

alpha
Numeric vector of values in the range [0, 1] for alpha transparency channel (0 means transparent and 1 means opaque).

rev
If TRUE (default), reverses the order of the colors in the color scale (compared to sequential_hcl).

begin
Number in the range of [0, 1] indicating to which point in the color scale the smallest data value should be mapped.

end
Number in the range of [0, 1] indicating to which point in the color scale the largest data value should be mapped.

na.value
Color to be used for missing data points.

guide
Type of legend. Use "coloursteps" for color bar with discrete steps.

aesthetics
The ggplot2 aesthetics to which this scale should be applied.

n_interp
Number of discrete colors that should be used to interpolate the binned color scale. 11 will work fine in most cases.

...
common binned scale parameters: 'name', 'breaks', 'labels', and 'limits'. See binned_scale for more details.

Details

If both a valid palette name and palette parameters are provided then the provided palette parameters overwrite the parameters in the named palette. This enables easy customization of named palettes.

Compared to sequential_hcl the ordering of the colors in the sequential ggplot2 scale are reversed by default (i.e., rev = TRUE) to be more consistent with ggplot2's own scales such as scale_color_fermenter. For most named palettes this leads to darker and more colorful colors for larger values on the scale. This is typically the better default on light/white backgrounds.

Examples

library("ggplot2")

# volcano plot
df <- data.frame(height = c(volcano), x = c(row(volcano)), y = c(col(volcano)))
ggplot(df, aes(x, y, fill = height)) + 
  geom_raster() + scale_fill_binned_sequential(palette = "Terrain", rev = FALSE) +
  coord_fixed(expand = FALSE)

[Package colorspace version 2.0-2 Index]