scale_colour_discrete_qualitative {colorspace}	R Documentation

HCL-Based Discrete Qualitative Color Scales for ggplot2

Description

Discrete ggplot2 color scales using the color palettes generated by qualitative_hcl.

Usage

scale_colour_discrete_qualitative(
  palette = NULL,
  c1 = NULL,
  l1 = NULL,
  h1 = NULL,
  h2 = NULL,
  alpha = 1,
  rev = FALSE,
  nmax = NULL,
  order = NULL,
  aesthetics = "colour",
  ...
)

scale_color_discrete_qualitative(
  palette = NULL,
  c1 = NULL,
  l1 = NULL,
  h1 = NULL,
  h2 = NULL,
  alpha = 1,
  rev = FALSE,
  nmax = NULL,
  order = NULL,
  aesthetics = "colour",
  ...
)

scale_fill_discrete_qualitative(..., aesthetics = "fill")

Arguments

palette
The name of the palette to be used. Run hcl_palettes(type = "qualitative") for available options.

c1
Chroma value, used for all colors in the scale.

l1
Luminance value, used for all colors in the scale.

h1
Beginning hue value.

h2
Ending hue value.

alpha
Numeric vector of values in the range [0, 1] for alpha transparency channel (0 means transparent and 1 means opaque).

rev
If TRUE, reverses the order of the colors in the color scale.

nmax
Maximum number of different colors the palette should contain. If not provided, is calculated automatically from the data.

order
Numeric vector listing the order in which the colors should be used. Default is 1:nmax.

aesthetics
The ggplot2 aesthetics to which this scale should be applied.

...
common discrete scale parameters: name, breaks, labels, na.value, limits and guide. See discrete_scale for more details.

Details

If both a valid palette name and palette parameters are provided then the provided palette parameters overwrite the parameters in the named palette. This enables easy customization of named palettes.

Examples

library("ggplot2")

# default colors
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() + theme_minimal() +
  scale_color_discrete_qualitative()
 
# color scale "Harmonic"
ggplot(iris, aes(Sepal.Length, fill = Species)) +
  geom_density(alpha = 0.7) + scale_fill_discrete_qualitative(palette = "Harmonic")

[Package colorspace version 2.0-2 Index]