whitepoint {colorspace}	R Documentation

Access or Modify the Whitepoint

Description

This function can be used to control the single global whitepoint that affects all color conversions within the package (that require a whitepoint, i.e., go through XYZ).

Usage

whitepoint(white, ...)

Arguments

white, ...
Either missing (to query the whitepoint) or NULL or a specification of the XYZ coordinates of the whitepoint (to set the whitepoint, see examples). NULL corresponds to CIE D65 with XYZ coordinates 95.047, 100.000, 108.883.

Value

whitepoint returns an XYZ color object for the whitepoint (invisibly in case a new whitepoint was set).

See Also

XYZ and color-class.

Examples

# query current whitepoint (D65 by default)
whitepoint()

# Illuminant E 
whitepoint(XYZ(100, 100, 100))

# equivalently
whitepoint(100, 100, 100)
whitepoint(c(100, 100, 100))
whitepoint(cbind(100, 100, 100))

whitepoint()

## reset
whitepoint(NULL)
whitepoint()

[Package colorspace version 2.0-2 Index]